JIRA tickets BMS-99, BMS-100, BMS-98 are all done now from this MR, thing to note is that the formatting may not be the way we want but is highly adjustable and can be JSONIFIED if needed by a library of sorts.
