+++ commands/saved_data.txt
@@ -1,0 +1,47 @@
+[
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 363.487
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 25.937
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 7.877
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 3.222
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 2.965
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 2.666
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 2.419
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 2.095
+  },
+  {
+    "numOfReactions": 0,
+    "response_image": null,
+    "timeToPost": 1.309
+  }
+]
